moves the widgets about news marquee into newsMarquee folder
