+++ lib/widgets/newsMarquee/marqueeWidget.dart
